requirements.txt
python-dotenv==0.20.0
python-telegram-bot==13.12
requests==2.27.1
google-cloud-dialogflow==2.14.0
